pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "openlinkedin"
version = "0.1.0"
description = "LinkedIn automation system with AI-powered content generation"
requires-python = ">=3.10"
dependencies = [
    "playwright>=1.40",
    "playwright-stealth>=1.0.6",
    "openai>=1.12",
    "anthropic>=0.18",
    "chromadb>=0.4",
    "sentence-transformers>=2.2",
    "apscheduler>=3.10,<4",
    "streamlit>=1.30",
    "pydantic>=2.5",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "tenacity>=8.2",
    "termcolor>=2.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
